Dlemon search engine
Building a complete search-solution using open source tool ‘Solr’, where
multilingual twitter data is collected using Twitter API and java.

In Part A of the project, data ingestion and indexing was done using
‘Solr’.
In Part B implementation of various Information Retrieval will be done
on target set of queries.
In Part C, we are building a search engine that runs on more than 5000
tweets using ‘Solr’